part two solution in progress

+++ src/Operand.java
@@ -1,5 +1,5 @@
 public class Operand {
-    public static Integer getOperandValue(Integer operand, Integer registerA,Integer registerB,Integer registerC){
+    public static long getOperandValue(long operand, long registerA,long registerB,long registerC){
         if (operand<=3||operand==7) return operand;
         else if (operand==4) return registerA;
         else if (operand==5) return registerB;
